Monitoring data about transmitters and logging can be sent to a Graphite server for graphical viewing.
git-svn-id: http://mireka.googlecode.com/svn/trunk@147 741d60a8-641f-389e-c7ec-093451e00b49

+++ src/mireka/transmission/queue/TransmitterSummaryMBean.java
@@ -1,17 +1,17 @@
 package mireka.transmission.queue;
 
 public interface TransmitterSummaryMBean {
-    int getMailTransactions();
+    long getMailTransactions();
 
-    int getSuccessfulMailTransactions();
+    long getSuccessfulMailTransactions();
 
-    int getFailures();
+    long getFailures();
 
-    int getPartialFailures();
+    long getPartialFailures();
 
-    int getPermanentFailures();
+    long getPermanentFailures();
 
-    int getTransientFailures();
+    long getTransientFailures();
 
     double getFailuresPercentage();
 
@@ -21,7 +21,7 @@ public interface TransmitterSummaryMBean {
 
     String getLastFailure();
 
-    int getErrors();
+    long getErrors();
 
     String getLastError();
 }
